+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Fix issues with `enum` schema classes (e.g. equality not working, `to_s` not working)
+- Changed assumption in the base schema class that it was a record (e.g. defining `to_h` instead of `as_json`). Moved record functionality to the record base class.
 - Fixed bug with previous release where the filename and constant names for `generate_namespace_folders` did not always match.
 # 1.14.6 - 2022-06-21